requirements.txt
Django==5.2.8
Pillow==11.3.0
django-ckeditor==6.7.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
